requirements.txt
datasets==1.17.0
nltk==3.5
pandas==1.3.5
Pillow==9.0.0
scikit-learn==0.23.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.0+cu111
torchvision==0.11.0+cu111
transformers==4.14.0
accelerate
